LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/with-types/at-rule-with-types-parsing-expected.txt


FAIL types: nOnE shoud result in [] assert_equals: expected "@view-transition { types: none; }" but got "@view-transition { }"
FAIL types: abc shoud result in ["abc"] assert_equals: expected "@view-transition { types: abc; }" but got "@view-transition { types: abc}}"
FAIL types: abc xyz shoud result in ["abc","xyz"] assert_equals: expected "@view-transition { types: abc xyz; }" but got "@view-transition { types: abc xyz}}"
FAIL types: abc\ xyz shoud result in ["abc xyz"] assert_equals: expected "@view-transition { types: abc\\ xyz; }" but got "@view-transition { types: abc xyz}}"
FAIL types:    abc 	xyz  shoud result in ["abc","xyz"] assert_equals: expected "@view-transition { types: abc xyz; }" but got "@view-transition { types: abc xyz}}"
PASS types: abc none shoud result in []
PASS types: abc, def shoud result in []
PASS types: dEfAuLt shoud result in []
PASS types: abc -ua-something shoud result in []
FAIL types: abc -uA-something shoud result in ["abc","-uA-something"] assert_equals: expected "@view-transition { types: abc -uA-something; }" but got "@view-transition { types: abc -uA-something}}"
FAIL types: abc -ok-something shoud result in ["abc","-ok-something"] assert_equals: expected "@view-transition { types: abc -ok-something; }" but got "@view-transition { types: abc -ok-something}}"
FAIL types: abc abc shoud result in ["abc","abc"] assert_equals: expected "@view-transition { types: abc abc; }" but got "@view-transition { types: abc abc}}"
FAIL types: abc ABC shoud result in ["abc","ABC"] assert_equals: expected "@view-transition { types: abc ABC; }" but got "@view-transition { types: abc ABC}}"
PASS types: 123 shoud result in []
PASS types: * shoud result in []
PASS types: *11 abc shoud result in []
